+++ runtime/src/main/scala/akka/grpc/javadsl/GrpcExceptionHandler.scala
@@ -7,9 +7,9 @@ package akka.grpc.javadsl
 import java.util.concurrent.CompletionException
 
 import akka.actor.ActorSystem
-import akka.grpc.{ GrpcProtocolNative, GrpcServiceException, Identity }
+import akka.grpc.{ GrpcServiceException, Identity }
 import akka.grpc.GrpcProtocol.GrpcProtocolWriter
-import akka.grpc.internal.{ GrpcResponseHelpers, MissingParameterException }
+import akka.grpc.internal.{ GrpcProtocolNative, GrpcResponseHelpers, MissingParameterException }
 import akka.http.javadsl.model.HttpResponse
 import akka.japi.{ Function => jFunction }
 import io.grpc.Status
